Rewrite tooltip display logic
This commit renames the `tooltipMessage` prop to `staticTooltipMessage` to make it explicit that it will be the same on every day (compared to the dynamic/default message which is different on each day).
This commit also introduces multiple locales by default (English and French) chosen by a new `lang` prop, and renames `i18n` prop to `customI18n`.